Abstract fetching of route information into its own module
